Move notes and scene_changes to project_garbage
The notes and scene_changes are not really useful outside of the IVTC DN
GUI (scene changes for frames might be, but since they mark fields it's
not that relevant to the output) so moving them to project_garbage makes
that more obvious.

Also add a version to the project file, allowing cleaner backwards
compatibility going forward.

Also fix the height of scene change rectangle.